requirements.txt
django==1.8.6
djangorestframework==3.3.2
psycopg2==2.6.1
pygments==2.0.2
django-filter==0.12.0
django-oauth-toolkit==0.9.0
django-cors-headers==1.1.0